feat(api): add ownership checks to course crud

+++ devcamper_api/src/feat/courses/course.controller.ts
@@ -3,6 +3,8 @@ import CourseR from "./course.repository.ts";
 import BootcampR from "../bootcamps/bootcamp.repository.ts";
 import asyncHandler from "../../middleware/asyncHandler.ts";
 import ErrorResponse from "../../utils/errorResponse.ts";
+import type {HydratedDocument} from "mongoose";
+import type {IUser} from "../users/user.model.ts";
 
 /**
  * @desc Get all courses
@@ -64,12 +66,25 @@ export const getCourse = asyncHandler(async (req: Request, res: Response, next: 
  * @access Private
  */
 export const createCourse = asyncHandler(async (req: Request, res: Response, next: NextFunction) => {
+    const user = res.locals.user as HydratedDocument<IUser>|null|undefined;
+    if(!user) {return;}
+
     const bootcamp = await BootcampR.fetchBootcamp(req.params.bootcampId);
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found`, 404));
     }
+    // check ownership
+    if(res.locals.user
+        && res.locals.user !== 'admin'
+        && res.locals.user._id.toString() !== bootcamp.user.toString()
+    ) {
+        return next(new ErrorResponse(`User is not authorized to add a course to this bootcamp`, 401));
+    }
 
-    const course = await CourseR.createCourse(req.body, bootcamp._id);
+    const course = await CourseR.createCourse({
+        ...req.body,
+        user: user._id,
+    }, bootcamp._id);
 
     res.status(201).json({
         success: true,
@@ -83,17 +98,26 @@ export const createCourse = asyncHandler(async (req: Request, res: Response, nex
  * @access Private
  */
 export const updateCourse = asyncHandler(async (req: Request, res: Response, next: NextFunction) => {
-    const course = await CourseR.updateCourse(req.body, req.params.id);
+    const course = await CourseR.fetchCourse(req.params.id);
 
     if (!course) {
         return next(new ErrorResponse(`Course not found`, 404));
     }
+    // check ownership
+    if(res.locals.user
+        && res.locals.user !== 'admin'
+        && res.locals.user._id.toString() !== course.user.toString()
+    ) {
+        return next(new ErrorResponse(`User is not authorized to update this course`, 401));
+    }
+
+    const updatedCourse = await CourseR.updateCourse(req.body, req.params.id);
 
     res
         .status(200)
         .json({
             success: true,
-            data: course
+            data: updatedCourse
         });
 })
 
@@ -107,6 +131,13 @@ export const deleteCourse = asyncHandler(async (req: Request, res: Response, nex
 
     if (!course) {
         return next(new ErrorResponse(`Course not found`, 404));
+    }
+    // check ownership
+    if(res.locals.user
+        && res.locals.user !== 'admin'
+        && res.locals.user._id.toString() !== course.user.toString()
+    ) {
+        return next(new ErrorResponse(`User is not authorized to delete this course`, 401));
     }
 
     await CourseR.deleteCourse(course._id);
